Refactor: 대시보드(메인)페이지 SSR로 변경
